+++ tests/browser/common_selectors.py
@@ -117,8 +117,8 @@ class DashboardContentsOnSuccess(SelectorsEnum):
 
 class DashboardReadingProgress(SelectorsEnum):
     YOUR_PROGRESS_CARD = Selector(By.ID, 'your-progress-card')
-    TOPICS_HEADERS = Selector(By.CSS_SELECTOR, 'h3[id^=topics-]')
-    TOPICS_READ_PROGRESS = Selector(By.CSS_SELECTOR, 'div[id^=topics-read-progress-]')
+    TOPICS_HEADERS = Selector(By.CSS_SELECTOR, '#your-progress-card h3[id^=topics-]')
+    TOPICS_READ_PROGRESS = Selector(By.CSS_SELECTOR, '#your-progress-card progress[id^=topics-read-]')
 
 
 class SignUpModal(SelectorsEnum):