+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 - GP2-315 - route to market backend integration
 
 ### Implemented enhancements
+- GP2-336 - Lesson progress on dashboard
 - GP2-386 - Lesson - hero image (FE) - Resize
 - no ticket - rename fiction components to fictional example
 - No ticket - wagtail upgrade 2.10